Revised to use CMake and new "64-bit-ready" Java JNI. Unfortunately, CMake builds non-relocatable project files, so I need write something to fix the Visual Studio projects so they can be used without CMake.

+++ portmidi/trunk/porttime/ptwinmm.c
@@ -64,7 +64,7 @@ PtTimestamp Pt_Time()
 }
 
 
-void Pt_Sleep(long duration)
+void Pt_Sleep(int32_t duration)
 {
     Sleep(duration);
 }
